# lib/api-zod/src/generated/types/previewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base44 to GitHub push tool API
 * OpenAPI spec version: 0.1.0
 */

export interface PreviewInput {
  base44AppId: string;
  base44ApiKey: string;
}