# lib/api-zod/src/generated/types/previewResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base44 to GitHub push tool API
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntry } from './fileEntry';

export interface PreviewResult {
  appName: string;
  files: FileEntry[];
}