# lib/api-zod/src/generated/types/pushInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base44 to GitHub push tool API
 * OpenAPI spec version: 0.1.0
 */

export interface PushInput {
  /** The Base44 App ID */
  base44AppId: string;
  /** The Base44 API Key */
  base44ApiKey: string;
  /** GitHub Personal Access Token */
  githubToken: string;
  /** GitHub repository owner (username or org) */
  githubOwner: string;
  /** GitHub repository name */
  githubRepo: string;
  /** Target branch (defaults to main) */
  branch?: string;
  /** Commit message */
  commitMessage?: string;
}