# lib/api-zod/src/generated/types/pushResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base44 to GitHub push tool API
 * OpenAPI spec version: 0.1.0
 */

export interface PushResult {
  success: boolean;
  filesCount: number;
  commitUrl: string;
  message: string;
}